# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Service API
  version: 1.0.0
  description: API contract for the inventory service MVP.
servers:
  - url: http://localhost:8080
    description: Local dev server
security:
  - SessionAuth: []
paths:
...
  /api/v1/equipment:
    get:
      summary: List equipment
      operationId: listEquipment
      responses:
        '200':
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipment'
    post:
      summary: Create equipment
      operationId: createEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreateRequest'
      responses:
        '201':
          description: Equipment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate asset name or serial number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/equipment/{id}:
    put:
      summary: Update equipment
      operationId: updateEquipment
      parameters:
        - $ref: '#/components/parameters/EquipmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentUpdateRequest'
      responses:
        '200':
          description: Equipment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete equipment
      operationId: deleteEquipment
      parameters:
        - $ref: '#/components/parameters/EquipmentId'
      responses:
        '200':
          description: Equipment deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/import:
    post:
      summary: Import equipment data from CSV
      operationId: importEquipment
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Import completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
  /api/v1/borrow-requests:
    get:
      summary: List borrow requests
      operationId: listBorrowRequests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: equipment_id
          in: query
          schema:
            type: integer
        - name: borrower_name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Borrow request list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BorrowRequest'
    post:
      summary: Create borrow request
      operationId: createBorrowRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowRequestCreateRequest'
      responses:
        '201':
          description: Borrow request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRequest'
        '404':
          description: Equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Equipment unavailable or already requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid borrow type, missing return date, or missing borrower name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden for non-user roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/borrow-requests/{id}:
    get:
      summary: Get borrow request by id
      operationId: getBorrowRequest
      parameters:
        - $ref: '#/components/parameters/BorrowRequestId'
      responses:
        '200':
          description: Borrow request found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRequest'
        '404':
          description: Borrow request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/borrow-requests/{id}/approve:
    put:
      summary: Approve borrow request
      operationId: approveBorrowRequest
      parameters:
        - $ref: '#/components/parameters/BorrowRequestId'
      responses:
        '200':
          description: Borrow request approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRequest'
        '404':
          description: Borrow request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Borrow request already approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/borrow-requests/{id}/reject:
    put:
      summary: Reject borrow request
      operationId: rejectBorrowRequest
      parameters:
        - $ref: '#/components/parameters/BorrowRequestId'
      responses:
        '200':
          description: Borrow request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRequest'
        '404':
          description: Borrow request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Borrow request not pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/borrow-requests/{id}/confirm-return:
    put:
      summary: Confirm return of borrow request
      operationId: confirmReturnBorrowRequest
      parameters:
        - $ref: '#/components/parameters/BorrowRequestId'
      responses:
        '200':
          description: Return confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowRequest'
        '404':
          description: Borrow request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Borrow request not pending return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    EquipmentId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    BorrowRequestId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    Equipment:
      type: object
      properties:
        id:
          type: integer

        type_id:
          type: integer

        type_name:
          type: string
          description: Equipment type name from equipment_types table

        product_name:
          type: string

        asset_name:
          type: string

        asset_tag:
          type: string
          nullable: true

        asset_serial_no:
          type: string
          nullable: true

        bar_code:
          type: string
          nullable: true

        vendor_name:
          type: string
          nullable: true

        location:
          type: string
          nullable: true

        assigned_to_department:
          type: string
          nullable: true

        site:
          type: string
          nullable: true

        asset_no:
          type: string
          nullable: true

        budget:
          type: string
          maxLength: 10
          nullable: true

        remark:
          type: string
          nullable: true

        username:
          type: string
          nullable: true
          description: UserName associated with the equipment record from inventory data

        req_no:
          type: string
          nullable: true

        status:
          type: string
          enum:
            - กำลังใช้งาน
            - ในคลัง
            - เสียหาย
            - เลิกใช้งาน

        created_at:
          type: string
          format: date-time

        updated_at:
          type: string
          format: date-time

      required:
        - id
        - type_id
        - type_name
        - product_name
        - asset_name
        - status
        - created_at
        - updated_at
    EquipmentCreateRequest:
      type: object
      properties:
        type_id:
          type: integer
          minimum: 1

        product_name:
          type: string
          minLength: 1
          pattern: '.*\S.*'

        asset_name:
          type: string
          minLength: 1
          pattern: '.*\S.*'

        asset_tag:
          type: string

        asset_serial_no:
          type: string

        bar_code:
          type: string

        vendor_name:
          type: string

        location:
          type: string

        assigned_to_department:
          type: string

        site:
          type: string

        asset_no:
          type: string

        budget:
          type: string
          maxLength: 10

        remark:
          type: string

        username:
          type: string
          description: UserName associated with the equipment record from inventory data

        req_no:
          type: string

      required:
        - type_id
        - product_name
        - asset_name
    EquipmentUpdateRequest:
      type: object
      properties:
        type_id:
          type: integer
          minimum: 1

        product_name:
          type: string
          minLength: 1
          pattern: '.*\S.*'

        asset_name:
          type: string
          minLength: 1
          pattern: '.*\S.*'

        asset_tag:
          type: string

        asset_serial_no:
          type: string

        bar_code:
          type: string

        vendor_name:
          type: string

        location:
          type: string

        assigned_to_department:
          type: string

        site:
          type: string

        asset_no:
          type: string

        budget:
          type: string
          maxLength: 10

        remark:
          type: string

        username:
          type: string
          description: UserName associated with the equipment record from inventory data

        req_no:
          type: string

        status:
          type: string
          enum:
            - กำลังใช้งาน
            - ในคลัง
            - เสียหาย
            - เลิกใช้งาน

      required:
        - type_id
        - product_name
        - asset_name
        - status
    ImportResult:
      type: object
      properties:
        total:
          type: integer
        success:
          type: integer
        failed:
          type: integer
      required:
        - total
        - success
        - failed
    BorrowRequest:
      properties:
        id: {type: integer}
        equipment_id: {type: integer}
        created_by_user_id: {type: integer}
        borrower_name: {type: string}
        borrow_type: {type: string}
        return_date: {type: string, format: date}
        status: {type: string}
        approved_by: {type: integer}
        returned_at: {type: string, format: date-time}
      required:
        - id
        - equipment_id
        - created_by_user_id
        - borrower_name
        - borrow_type
        - status
    BorrowRequestCreateRequest:
      properties:
        equipment_id: {type: integer}
        borrow_type: {type: string}
        return_date: {type: string, format: date}
        borrower_name: {type: string}
      required:
        - equipment_id
        - borrow_type
        - borrower_name
...